// src/components/common/Form/enums.ts
export enum SchemaField {
  InputText = 'InputText',
  InputNumber = 'InputNumber',
  InputPassword = 'InputPassword',
  InputCurrency = 'InputCurrency',
  InputCheckbox = 'InputCheckbox',
  InputRadio = 'InputRadio',
  InputSwitch = 'InputSwitch',
  SwitchButtons = 'SwitchButtons',
  InputSelect = 'InputSelect',
  InputMultiSelect = 'InputMultiSelect',
  InputAutoComplete = 'InputAutoComplete',
  InputAsyncAutoComplete = 'InputAsyncAutoComplete',
  InputDate = 'InputDate',
  InputPhoneNumber = 'InputPhoneNumber',
  InputTabs = 'InputTabs',
  InputTextEditor = 'InputTextEditor',
  InputHidden = 'InputHidden',
  InputIBAN = 'InputIBAN',
  CustomComponent = 'CustomComponent',
  InputNumericMask = 'InputNumericMask',
}

export enum SchemaInputType {
  file = 'file',
  checkbox = 'checkbox',
  text = 'text',
  url = 'url',
  email = 'email',
  date = 'date',
  textarea = 'textarea',
  number = 'number',
  password = 'password',
  texteditor = 'texteditor',
  radio = 'radio',
}